refactor batch_track and create a new file for cli
